added confirmation and refresh to delete

+++ client/src/helpers/requests.js
@@ -3,17 +3,19 @@ export async function deleteLink(linkId) {
     
     const link = `/links/${linkId}`
     console.log(link)
-    const response = await fetch(link, {
-        method: 'DELETE',
-        headers: {
-        'Content-Type': 'application/json'
-        },
-    })
-    if (!response.ok) {
-        throw new Error('Could not DELETE link :(')
+    if (window.confirm("Do you want to delete the link?")) {
+        const response = await fetch(link, {
+            method: 'DELETE',
+            headers: {
+            'Content-Type': 'application/json'
+            },
+        }).then(() => {
+            window.location.reload()
+        })
+        if (!response.ok) {
+            throw new Error('Could not DELETE link :(')
+        }
     }
-
-    //fetch    
 }
 
 export async function fetchList() {
